Split home, about, contact to their own module
