+++ src/index.js
@@ -2,7 +2,7 @@ import './bulma.min.css';
 import './style.css';
 import renderFooter from './renderFooter';
 import renderNavbar from './renderNavbar';
-import { renderMenuSection } from './renders';
+import renderMenuSection from './renderHome';
 
 const renderPage = () => {
   const contentDiv = document.querySelector('#content');
